+++ js/data.js
@@ -1,0 +1,14 @@
+const TITLES = ['Объявление-1', 'Объявление-2', 'Объявление-3', 'Объявление-4'];
+const ACCOMMODATIONS = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+const CHECKINS = ['12:00', '13:00', '14:00'];
+const CHECKOUTS = ['12:00', '13:00', '14:00'];
+const FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+const DESCRIPTIONS = ['Просторное', 'Удобное', 'Уютное', 'Комфортное'];
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+
+export { TITLES, ACCOMMODATIONS, CHECKINS, CHECKOUTS, FEATURES, DESCRIPTIONS, PHOTOS };
